+++ src/app/guards/dashboard-auth.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CanActivate } from '@angular/router';
+import { CanActivate, Router } from '@angular/router';
 import { UserService } from '../services/user.service';
 import { GuildService } from '../services/guild.service';
 
@@ -9,14 +9,18 @@ import { GuildService } from '../services/guild.service';
 export class DashboardAuthGuard implements CanActivate {
   constructor(
     private userService: UserService,
-    private guildService: GuildService) {}
+    private guildService: GuildService,
+    private router: Router) {}
 
   async canActivate() {
     if (!this.userService.user)
       await this.userService.updateUser();
     if (!this.userService.savedUser)
       await this.userService.updateSavedUser();
-      
-    return Boolean(this.userService.user);
+    
+    const canActivate = Boolean(this.userService.user);
+    if (!canActivate)
+      this.router.navigate(['/login'])
+    return canActivate;
   }  
 }