+++ src/app/guards/leaderboard-auth.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { GuildService } from '../services/guild.service';
 import { UserService } from '../services/user.service';
 
@@ -9,7 +9,8 @@ import { UserService } from '../services/user.service';
 export class LeaderboardAuthGuard implements CanActivate {
   constructor(
     private guildService: GuildService,
-    private userService: UserService) {}
+    private userService: UserService,
+    private router: Router) {}
 
   async canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
     const id = next.paramMap.get('id');
@@ -17,7 +18,11 @@ export class LeaderboardAuthGuard implements CanActivate {
     if (guildConfig?.settings.privateLeaderboard) {
       const members = await this.guildService.getMembers(id);
       return members.some(m => m.id === this.userService.user.id);
+    
     }
-    return !guildConfig || !guildConfig.settings.privateLeaderboard;
+    const canActivate = guildConfig && !guildConfig.settings.privateLeaderboard;
+    if (!canActivate)
+      this.router.navigate(['/404'])
+    return canActivate;
   }
 }